Strix 2.1.2
Feats
1. `PopularityUtil` 更新.
2. `RedisUtil` 更新. 增加 `ZSET` 相关方法, 增加 `isType` 方法.

Fixes

+++ src/main/java/cn/projectan/strix/model/request/system/tool/popularity/PopularityDataUpdateReq.java
@@ -17,6 +17,6 @@ public class PopularityDataUpdateReq {
     @NotNull(groups = {InsertGroup.class, UpdateGroup.class}, message = "数值不可为空")
     @Min(groups = {InsertGroup.class, UpdateGroup.class}, value = 0, message = "数值超出范围")
     @UpdateField
-    private Integer originalValue;
+    private Long originalValue;
 
 }
